reorg edge packages
