+++ nvflare/edge/__init__.py
@@ -11,21 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from nvflare.fuel.f3.cellnet.defs import ReturnCode as CellReturnCode
-
-
-class Status(CellReturnCode):
-    NO_TASK = "no_task"
-    NO_JOB = "no_job"
-
-
-class EdgeProtoKey:
-    STATUS = "status"
-    DATA = "data"
-
-
-class EdgeContextKey:
-    JOB_ID = "__edge_job_id__"
-    EDGE_CAPABILITIES = "__edge_capabilities__"
-    REQUEST_FROM_EDGE = "__request_from_edge__"
-    REPLY_TO_EDGE = "__reply_to_edge__"
